+++ src/axom/inlet/tests/inlet_object.cpp
@@ -14,20 +14,21 @@
 
 #include "axom/sidre.hpp"
 
-#include "axom/inlet/LuaReader.hpp"
 #include "axom/inlet/Inlet.hpp"
+#include "axom/inlet/tests/inlet_test_utils.hpp"
 
 using axom::inlet::Inlet;
 using axom::inlet::InletType;
 using axom::inlet::LuaReader;
 using axom::sidre::DataStore;
 
+template <typename InletReader>
 Inlet createBasicInlet(DataStore* ds,
                        const std::string& luaString,
                        bool enableDocs = true)
 {
-  auto lr = std::make_unique<LuaReader>();
-  lr->parseString(luaString);
+  auto lr = std::make_unique<InletReader>();
+  lr->parseString(axom::inlet::detail::fromLuaTo<InletReader>(luaString));
 
   return Inlet(std::move(lr), ds->getRoot(), enableDocs);
 }
@@ -53,11 +54,17 @@ struct FromInlet<Foo>
   }
 };
 
-TEST(inlet_object, simple_struct_by_value)
+template <typename InletReader>
+class inlet_object : public ::testing::Test
+{ };
+
+TYPED_TEST_SUITE(inlet_object, axom::inlet::detail::ReaderTypes);
+
+TYPED_TEST(inlet_object, simple_struct_by_value)
 {
   std::string testString = "foo = { bar = true; baz = false }";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
   // Define schema
   inlet.addBool("foo/bar", "bar's description");
@@ -70,53 +77,53 @@ TEST(inlet_object, simple_struct_by_value)
   EXPECT_FALSE(foo.baz);
 }
 
-TEST(inlet_object, simple_array_of_struct_by_value)
+TYPED_TEST(inlet_object, simple_array_of_struct_by_value)
 {
   std::string testString =
-    "foo = { [4] = { bar = true; baz = false}, "
-    "        [7] = { bar = false; baz = true} }";
+    "foo = { [0] = { bar = true; baz = false}, "
+    "        [1] = { bar = false; baz = true} }";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.getGlobalTable().addGenericArray("foo");
+  auto& arr_table = inlet.addGenericArray("foo");
 
   arr_table.addBool("bar", "bar's description");
   arr_table.addBool("baz", "baz's description");
-  std::unordered_map<int, Foo> expected_foos = {{4, {true, false}},
-                                                {7, {false, true}}};
-  std::unordered_map<int, Foo> foos;
-  EXPECT_TRUE(arr_table.getArray(foos));
+  std::unordered_map<int, Foo> expected_foos = {{0, {true, false}},
+                                                {1, {false, true}}};
+  auto foos = inlet["foo"].get<std::unordered_map<int, Foo>>();
   EXPECT_EQ(foos, expected_foos);
 }
 
-TEST(inlet_object, simple_array_of_struct_implicit_idx)
+TYPED_TEST(inlet_object, simple_array_of_struct_implicit_idx)
 {
   std::string testString =
     "foo = { { bar = true; baz = false}, "
     "        { bar = false; baz = true} }";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.getGlobalTable().addGenericArray("foo");
+  auto& arr_table = inlet.addGenericArray("foo");
 
   arr_table.addBool("bar", "bar's description");
   arr_table.addBool("baz", "baz's description");
-  std::unordered_map<int, Foo> expected_foos = {{1, {true, false}},
-                                                {2, {false, true}}};
-  std::unordered_map<int, Foo> foos;
-  EXPECT_TRUE(arr_table.getArray(foos));
+  // Lua is 1-indexed
+  const int base_idx = TypeParam::baseIndex;
+  std::unordered_map<int, Foo> expected_foos = {{base_idx, {true, false}},
+                                                {base_idx + 1, {false, true}}};
+  auto foos = inlet["foo"].get<std::unordered_map<int, Foo>>();
   EXPECT_EQ(foos, expected_foos);
 }
 
-TEST(inlet_object, simple_array_of_struct_verify_optional)
+TYPED_TEST(inlet_object, simple_array_of_struct_verify_optional)
 {
   std::string testString =
     "foo = { [4] = { bar = true;}, "
     "        [7] = { bar = false;} }";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.getGlobalTable().addGenericArray("foo");
+  auto& arr_table = inlet.addGenericArray("foo");
 
   arr_table.addBool("bar", "bar's description").required(true);
   arr_table.addBool("baz", "baz's description").required(false);
@@ -124,15 +131,15 @@ TEST(inlet_object, simple_array_of_struct_verify_optional)
   EXPECT_TRUE(inlet.verify());
 }
 
-TEST(inlet_object, simple_array_of_struct_verify_reqd)
+TYPED_TEST(inlet_object, simple_array_of_struct_verify_reqd)
 {
   std::string testString =
     "foo = { [4] = { bar = true;}, "
     "        [7] = { bar = false;} }";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.getGlobalTable().addGenericArray("foo");
+  auto& arr_table = inlet.addGenericArray("foo");
 
   arr_table.addBool("bar", "bar's description").required(true);
   arr_table.addBool("baz", "baz's description").required(true);
@@ -156,21 +163,22 @@ struct FromInlet<FooWithArray>
   }
 };
 
-TEST(inlet_object, array_of_struct_containing_array)
+TYPED_TEST(inlet_object, array_of_struct_containing_array)
 {
   std::string testString =
-    "foo = { [4] = { arr = { [1] = 3 }; }, "
-    "        [7] = { arr = { [6] = 2 }; } }";
+    "foo = { [0] = { arr = { [0] = 3 }; }, "
+    "        [1] = { arr = { [0] = 2 }; } }";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
-  auto& arr_table = inlet.getGlobalTable().addGenericArray("foo");
+  auto& arr_table = inlet.addGenericArray("foo");
 
   arr_table.addIntArray("arr", "arr's description");
-  std::unordered_map<int, FooWithArray> expected_foos = {{4, {{{1, 3}}}},
-                                                         {7, {{{6, 2}}}}};
+  // Contiguous indexing for generality
+  std::unordered_map<int, FooWithArray> expected_foos = {{0, {{{0, 3}}}},
+                                                         {1, {{{0, 2}}}}};
   std::unordered_map<int, FooWithArray> foos_with_arr;
-  EXPECT_TRUE(arr_table.getArray(foos_with_arr));
+  foos_with_arr = inlet["foo"].get<std::unordered_map<int, FooWithArray>>();
   EXPECT_EQ(foos_with_arr, expected_foos);
 }
 
@@ -194,11 +202,11 @@ struct FromInlet<MoveOnlyFoo>
   }
 };
 
-TEST(inlet_object, simple_moveonly_struct_by_value)
+TYPED_TEST(inlet_object, simple_moveonly_struct_by_value)
 {
   std::string testString = "foo = { bar = true; baz = false }";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
   // Define schema
   // Check for existing fields
@@ -211,11 +219,11 @@ TEST(inlet_object, simple_moveonly_struct_by_value)
   EXPECT_FALSE(foo.baz);
 }
 
-TEST(inlet_object, simple_value_from_bracket)
+TYPED_TEST(inlet_object, simple_value_from_bracket)
 {
   std::string testString = "foo = true";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
   // Define schema
   // Check for existing fields
@@ -225,11 +233,11 @@ TEST(inlet_object, simple_value_from_bracket)
   EXPECT_TRUE(foo);
 }
 
-TEST(inlet_object, simple_struct_from_bracket)
+TYPED_TEST(inlet_object, simple_struct_from_bracket)
 {
   std::string testString = "foo = { bar = true; baz = false }";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
   // Define schema
   // Check for existing fields
@@ -242,11 +250,11 @@ TEST(inlet_object, simple_struct_from_bracket)
   EXPECT_FALSE(foo.baz);
 }
 
-TEST(inlet_object, contains_from_table)
+TYPED_TEST(inlet_object, contains_from_table)
 {
   std::string testString = "foo = { bar = true; baz = false }";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
   // Define schema
   // Check for existing fields
@@ -262,11 +270,11 @@ TEST(inlet_object, contains_from_table)
   EXPECT_TRUE(foo_table.contains("baz"));
 }
 
-TEST(inlet_object, contains_from_bracket)
+TYPED_TEST(inlet_object, contains_from_bracket)
 {
   std::string testString = "foo = { bar = true; baz = false }";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
   // Define schema
   // Check for existing fields
@@ -278,29 +286,29 @@ TEST(inlet_object, contains_from_bracket)
   EXPECT_TRUE(inlet["foo"].contains("baz"));
 }
 
-TEST(inlet_object, array_from_bracket)
+TYPED_TEST(inlet_object, array_from_bracket)
 {
   DataStore ds;
   std::string testString =
-    "luaArrays = { arr1 = { [1] = 4}, "
-    "              arr2 = {[4] = true, [8] = false}, "
-    "              arr3 = {[33] = 'hello', [2] = 'bye'}, "
-    "              arr4 = { [12] = 2.4 } }";
-  auto inlet = createBasicInlet(&ds, testString);
+    "luaArrays = { arr1 = { [0] = 4}, "
+    "              arr2 = {[0] = true, [1] = false}, "
+    "              arr3 = {[0] = 'hello', [1] = 'bye'}, "
+    "              arr4 = { [0] = 2.4 } }";
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
   std::unordered_map<int, int> intMap;
   std::unordered_map<int, bool> boolMap;
   std::unordered_map<int, std::string> strMap;
   std::unordered_map<int, double> doubleMap;
-  inlet.getGlobalTable().addIntArray("luaArrays/arr1");
-  inlet.getGlobalTable().addBoolArray("luaArrays/arr2");
-  inlet.getGlobalTable().addStringArray("luaArrays/arr3");
-  inlet.getGlobalTable().addDoubleArray("luaArrays/arr4");
+  inlet.addIntArray("luaArrays/arr1");
+  inlet.addBoolArray("luaArrays/arr2");
+  inlet.addStringArray("luaArrays/arr3");
+  inlet.addDoubleArray("luaArrays/arr4");
 
-  std::unordered_map<int, int> expectedInts {{1, 4}};
-  std::unordered_map<int, bool> expectedBools {{4, true}, {8, false}};
-  std::unordered_map<int, std::string> expectedStrs {{33, "hello"}, {2, "bye"}};
-  std::unordered_map<int, double> expectedDoubles {{12, 2.4}};
+  std::unordered_map<int, int> expectedInts {{0, 4}};
+  std::unordered_map<int, bool> expectedBools {{0, true}, {1, false}};
+  std::unordered_map<int, std::string> expectedStrs {{0, "hello"}, {1, "bye"}};
+  std::unordered_map<int, double> expectedDoubles {{0, 2.4}};
 
   intMap = inlet["luaArrays/arr1"].get<std::unordered_map<int, int>>();
   EXPECT_EQ(intMap, expectedInts);
@@ -315,12 +323,12 @@ TEST(inlet_object, array_from_bracket)
   EXPECT_EQ(doubleMap, expectedDoubles);
 }
 
-TEST(inlet_object, primitive_type_checks)
+TYPED_TEST(inlet_object, primitive_type_checks)
 {
   std::string testString =
     " bar = true; baz = 12; quux = 2.5; corge = 'hello' ";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
   // Define schema
   // Check for existing fields
@@ -349,14 +357,14 @@ TEST(inlet_object, primitive_type_checks)
   EXPECT_EQ(corge, "hello");
 }
 
-TEST(inlet_object, composite_type_checks)
+TYPED_TEST(inlet_object, composite_type_checks)
 {
   std::string testString =
     "luaArrays = { arr1 = { [1] = 4}, "
     "              arr2 = {[4] = true, [8] = false} }; "
     "foo = { bar = true; baz = false }";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
   // Define schema
   // Check for existing fields
@@ -364,16 +372,16 @@ TEST(inlet_object, composite_type_checks)
 
   inlet.addBool("foo/baz", "baz's description");
 
-  inlet.getGlobalTable().addIntArray("luaArrays/arr1");
-  inlet.getGlobalTable().addBoolArray("luaArrays/arr2");
+  inlet.addIntArray("luaArrays/arr1");
+  inlet.addBoolArray("luaArrays/arr2");
 
   auto arr_table = inlet["luaArrays"];
   // The table containing the two arrays is not an array, but an object
   EXPECT_EQ(arr_table.type(), InletType::Object);
 
   // But the things it contains are arrays
-  EXPECT_EQ(arr_table["arr1"].type(), InletType::Array);
-  EXPECT_EQ(arr_table["arr2"].type(), InletType::Array);
+  EXPECT_EQ(arr_table["arr1"].type(), InletType::Container);
+  EXPECT_EQ(arr_table["arr2"].type(), InletType::Container);
 
   auto foo_table = inlet["foo"];
   // Similarly, the table containing the two bools is an object
@@ -384,13 +392,13 @@ TEST(inlet_object, composite_type_checks)
   EXPECT_EQ(foo_table["baz"].type(), InletType::Bool);
 }
 
-TEST(inlet_object, implicit_conversion_primitives)
+TYPED_TEST(inlet_object, implicit_conversion_primitives)
 {
   std::string testString =
-    " bar = true; baz = 12; quux = 2.5; corge = 'hello'; arr = { [1] = 4, [2] "
-    "= 6, [7] = 10}";
+    " bar = true; baz = 12; quux = 2.5; corge = 'hello'; arr = { [0] = 4, [1] "
+    "= 6, [2] = 10}";
   DataStore ds;
-  auto inlet = createBasicInlet(&ds, testString);
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
 
   // Define schema
   inlet.addBool("bar", "bar's description");
@@ -398,7 +406,7 @@ TEST(inlet_object, implicit_conversion_primitives)
   inlet.addDouble("quux", "quux's description");
   inlet.addString("corge", "corge's description");
 
-  inlet.getGlobalTable().addIntArray("arr");
+  inlet.addIntArray("arr");
 
   // Attempt both construction and assignment
   bool bar = inlet["bar"];
@@ -421,12 +429,260 @@ TEST(inlet_object, implicit_conversion_primitives)
   corge = inlet["corge"];
   EXPECT_EQ(corge, "hello");
 
-  std::unordered_map<int, int> expected_arr {{1, 4}, {2, 6}, {7, 10}};
+  std::unordered_map<int, int> expected_arr {{0, 4}, {1, 6}, {2, 10}};
   std::unordered_map<int, int> arr = inlet["arr"];
   EXPECT_EQ(arr, expected_arr);
   arr = inlet["arr"];
   EXPECT_EQ(arr, expected_arr);
 }
+
+template <typename InletReader>
+class inlet_object_dict : public ::testing::Test
+{ };
+
+TYPED_TEST_SUITE(inlet_object_dict, axom::inlet::detail::ReaderTypes);
+
+TYPED_TEST(inlet_object_dict, basic_dicts)
+{
+  std::string testString = "foo = { ['key1'] = 4, ['key3'] = 6, ['key2'] = 10}";
+  DataStore ds;
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
+
+  inlet.addIntDictionary("foo", "foo's description");
+  std::unordered_map<std::string, int> dict = inlet["foo"];
+  std::unordered_map<std::string, int> correct = {{"key1", 4},
+                                                  {"key3", 6},
+                                                  {"key2", 10}};
+  EXPECT_EQ(dict, correct);
+}
+
+TYPED_TEST(inlet_object_dict, simple_dict_of_struct_by_value)
+{
+  std::string testString =
+    "foo = { ['key1'] = { bar = true; baz = false}, "
+    "        ['key2'] = { bar = false; baz = true} }";
+  DataStore ds;
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
+
+  auto& dict_table = inlet.addGenericDictionary("foo");
+
+  dict_table.addBool("bar", "bar's description");
+  dict_table.addBool("baz", "baz's description");
+  std::unordered_map<std::string, Foo> expected_foos = {{"key1", {true, false}},
+                                                        {"key2", {false, true}}};
+  std::unordered_map<std::string, Foo> foos;
+  foos = inlet["foo"].get<std::unordered_map<std::string, Foo>>();
+  EXPECT_EQ(foos, expected_foos);
+}
+
+struct FooWithDict
+{
+  std::unordered_map<std::string, int> arr;
+  bool operator==(const FooWithDict& other) const { return arr == other.arr; }
+};
+
+template <>
+struct FromInlet<FooWithDict>
+{
+  FooWithDict operator()(const axom::inlet::Table& base)
+  {
+    FooWithDict f = {base["arr"]};
+    return f;
+  }
+};
+
+TYPED_TEST(inlet_object_dict, dict_of_struct_containing_dict)
+{
+  std::string testString =
+    "foo = { ['key3'] = { arr = { ['key1'] = 3 }; }, "
+    "        ['key4'] = { arr = { ['key2'] = 2 }; } }";
+  DataStore ds;
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
+
+  auto& dict_table = inlet.addGenericDictionary("foo");
+
+  dict_table.addIntDictionary("arr", "arr's description");
+  std::unordered_map<std::string, FooWithDict> expected_foos = {
+    {"key3", {{{"key1", 3}}}},
+    {"key4", {{{"key2", 2}}}}};
+  std::unordered_map<std::string, FooWithDict> foos_with_dict;
+  foos_with_dict =
+    inlet["foo"].get<std::unordered_map<std::string, FooWithDict>>();
+  EXPECT_EQ(foos_with_dict, expected_foos);
+}
+
+TYPED_TEST(inlet_object_dict, dict_of_struct_containing_array)
+{
+  std::string testString =
+    "foo = { ['key3'] = { arr = { [0] = 3 }; }, "
+    "        ['key4'] = { arr = { [0] = 2 }; } }";
+  DataStore ds;
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
+
+  auto& dict_table = inlet.addGenericDictionary("foo");
+
+  dict_table.addIntArray("arr", "arr's description");
+  std::unordered_map<std::string, FooWithArray> expected_foos = {
+    {"key3", {{{0, 3}}}},
+    {"key4", {{{0, 2}}}}};
+  std::unordered_map<std::string, FooWithArray> foos_with_array;
+  foos_with_array =
+    inlet["foo"].get<std::unordered_map<std::string, FooWithArray>>();
+  EXPECT_EQ(foos_with_array, expected_foos);
+}
+
+TYPED_TEST(inlet_object_dict, array_of_struct_containing_dict)
+{
+  std::string testString =
+    "foo = { [0] = { arr = { ['key1'] = 3 }; }, "
+    "        [1] = { arr = { ['key2'] = 2 }; } }";
+  DataStore ds;
+  Inlet inlet = createBasicInlet<TypeParam>(&ds, testString);
+
+  auto& arr_table = inlet.addGenericArray("foo");
+
+  arr_table.addIntDictionary("arr", "arr's description");
+  std::unordered_map<int, FooWithDict> expected_foos = {{0, {{{"key1", 3}}}},
+                                                        {1, {{{"key2", 2}}}}};
+  std::unordered_map<int, FooWithDict> foos_with_dict;
+  foos_with_dict = inlet["foo"].get<std::unordered_map<int, FooWithDict>>();
+  EXPECT_EQ(foos_with_dict, expected_foos);
+}
+
+/*
+FIXME: These are currently error conditions.  If these should be supported
+or handled differently these tests can be re-enabled.
+
+TEST(inlet_dict, mixed_keys)
+{
+  std::string testString = "foo = { ['key1'] = 4, [1] = 6 }";
+  DataStore ds;
+  Inlet inlet = createBasicInlet(&ds, testString);
+
+  inlet.addIntDictionary("foo", "foo's description");
+  std::unordered_map<std::string, int> dict = inlet["foo"];
+  std::unordered_map<std::string, int> correct_dict = {{"key1", 4}};
+  EXPECT_EQ(dict, correct_dict);
+
+  inlet.addIntArray("foo", "foo's description");
+  std::unordered_map<int, int> array = inlet["foo"];
+  std::unordered_map<int, int> correct_array = {{1, 6}};
+  EXPECT_EQ(array, correct_array);
+}
+
+TEST(inlet_dict, key_with_slash)
+{
+  std::string testString =
+    "foo = { ['key1/subkey1'] = 4, ['key3'] = 6, ['key2'] = 10}";
+  DataStore ds;
+  Inlet inlet = createBasicInlet(&ds, testString);
+
+  inlet.addIntDictionary("foo", "foo's description");
+  std::unordered_map<std::string, int> dict = inlet["foo"];
+  std::unordered_map<std::string, int> correct = {{"key1/subkey1", 4},
+                                                  {"key3", 6},
+                                                  {"key2", 10}};
+  EXPECT_EQ(dict, correct);
+}
+*/
+
+// Noncontiguous array tests that are only valid in Lua
+#ifdef AXOM_USE_SOL
+
+TEST(inlet_object_lua, array_of_struct_containing_array)
+{
+  std::string testString =
+    "foo = { [4] = { arr = { [1] = 3 }; }, "
+    "        [7] = { arr = { [6] = 2 }; } }";
+  DataStore ds;
+  Inlet inlet = createBasicInlet<axom::inlet::LuaReader>(&ds, testString);
+
+  auto& arr_table = inlet.addGenericArray("foo");
+
+  arr_table.addIntArray("arr", "arr's description");
+  std::unordered_map<int, FooWithArray> expected_foos = {{4, {{{1, 3}}}},
+                                                         {7, {{{6, 2}}}}};
+  std::unordered_map<int, FooWithArray> foos_with_arr;
+  foos_with_arr = inlet["foo"].get<std::unordered_map<int, FooWithArray>>();
+  EXPECT_EQ(foos_with_arr, expected_foos);
+}
+
+TEST(inlet_object_lua, array_from_bracket)
+{
+  DataStore ds;
+  std::string testString =
+    "luaArrays = { arr1 = { [1] = 4}, "
+    "              arr2 = {[4] = true, [8] = false}, "
+    "              arr3 = {[33] = 'hello', [2] = 'bye'}, "
+    "              arr4 = { [12] = 2.4 } }";
+  Inlet inlet = createBasicInlet<axom::inlet::LuaReader>(&ds, testString);
+
+  std::unordered_map<int, int> intMap;
+  std::unordered_map<int, bool> boolMap;
+  std::unordered_map<int, std::string> strMap;
+  std::unordered_map<int, double> doubleMap;
+  inlet.addIntArray("luaArrays/arr1");
+  inlet.addBoolArray("luaArrays/arr2");
+  inlet.addStringArray("luaArrays/arr3");
+  inlet.addDoubleArray("luaArrays/arr4");
+
+  std::unordered_map<int, int> expectedInts {{1, 4}};
+  std::unordered_map<int, bool> expectedBools {{4, true}, {8, false}};
+  std::unordered_map<int, std::string> expectedStrs {{33, "hello"}, {2, "bye"}};
+  std::unordered_map<int, double> expectedDoubles {{12, 2.4}};
+
+  intMap = inlet["luaArrays/arr1"].get<std::unordered_map<int, int>>();
+  EXPECT_EQ(intMap, expectedInts);
+
+  boolMap = inlet["luaArrays/arr2"].get<std::unordered_map<int, bool>>();
+  EXPECT_EQ(boolMap, expectedBools);
+
+  strMap = inlet["luaArrays/arr3"].get<std::unordered_map<int, std::string>>();
+  EXPECT_EQ(strMap, expectedStrs);
+
+  doubleMap = inlet["luaArrays/arr4"].get<std::unordered_map<int, double>>();
+  EXPECT_EQ(doubleMap, expectedDoubles);
+}
+
+TEST(inlet_object_lua_dict, dict_of_struct_containing_array)
+{
+  std::string testString =
+    "foo = { ['key3'] = { arr = { [1] = 3 }; }, "
+    "        ['key4'] = { arr = { [6] = 2 }; } }";
+  DataStore ds;
+  Inlet inlet = createBasicInlet<axom::inlet::LuaReader>(&ds, testString);
+
+  auto& dict_table = inlet.addGenericDictionary("foo");
+
+  dict_table.addIntArray("arr", "arr's description");
+  std::unordered_map<std::string, FooWithArray> expected_foos = {
+    {"key3", {{{1, 3}}}},
+    {"key4", {{{6, 2}}}}};
+  std::unordered_map<std::string, FooWithArray> foos_with_array;
+  foos_with_array =
+    inlet["foo"].get<std::unordered_map<std::string, FooWithArray>>();
+  EXPECT_EQ(foos_with_array, expected_foos);
+}
+
+TEST(inlet_object_lua_dict, array_of_struct_containing_dict)
+{
+  std::string testString =
+    "foo = { [7] = { arr = { ['key1'] = 3 }; }, "
+    "        [4] = { arr = { ['key2'] = 2 }; } }";
+  DataStore ds;
+  Inlet inlet = createBasicInlet<axom::inlet::LuaReader>(&ds, testString);
+
+  auto& arr_table = inlet.addGenericArray("foo");
+
+  arr_table.addIntDictionary("arr", "arr's description");
+  std::unordered_map<int, FooWithDict> expected_foos = {{7, {{{"key1", 3}}}},
+                                                        {4, {{{"key2", 2}}}}};
+  std::unordered_map<int, FooWithDict> foos_with_dict;
+  foos_with_dict = inlet["foo"].get<std::unordered_map<int, FooWithDict>>();
+  EXPECT_EQ(foos_with_dict, expected_foos);
+}
+
+#endif
 
 //------------------------------------------------------------------------------
 #include "axom/slic/core/UnitTestLogger.hpp"